requirements.txt
streamlit
openai
python-dotenv
pandas
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
